Finally fix annoying make_heap bug.

+++ umap/utils.py
@@ -157,7 +157,7 @@ def make_heap(n_points, size):
     heap: An ndarray suitable for passing to other numba enabled heap functions.
     """
     result = np.zeros(
-        (np.int32(3), np.int32(n_points), np.int32(size)), dtype=np.float64
+        (np.int64(3), np.int64(n_points), np.int64(size)), dtype=np.float64
     )
     result[0] = -1
     result[1] = np.infty
